requirements.txt
aiohttp==3.8.5
pycups==2.0.1
requests==2.31.0
python-slugify==8.0.1
PyYAML==6.0.1